+++ mathOperations/Subtraction.kt
@@ -1,5 +1,7 @@
 package calculator.mathOperations
 
+import java.math.BigDecimal
+
 object Subtraction : Runner {
-    override fun run(firstNum: Double, secondNum: Double) = firstNum - secondNum
+    override fun run(firstNum: BigDecimal, secondNum: BigDecimal) = firstNum - secondNum
 }
